Fix mastery badge reset

+++ app/Platform/Actions/RevalidateAchievementSetBadgeEligibility.php
@@ -6,7 +6,6 @@ namespace App\Platform\Actions;
 
 use App\Community\Enums\ActivityType;
 use App\Community\Enums\AwardType;
-use App\Platform\Enums\AchievementFlag;
 use App\Platform\Enums\UnlockMode;
 use App\Platform\Events\PlayerBadgeAwarded;
 use App\Platform\Events\PlayerBadgeLost;
@@ -14,7 +13,6 @@ use App\Platform\Events\PlayerGameBeaten;
 use App\Platform\Events\PlayerGameCompleted;
 use App\Platform\Models\PlayerGame;
 use Carbon\Carbon;
-use Illuminate\Support\Facades\DB;
 
 class RevalidateAchievementSetBadgeEligibility
 {
@@ -35,8 +33,8 @@ class RevalidateAchievementSetBadgeEligibility
         $badge = $playerGame->user->playerBadges()
             ->where('AwardType', AwardType::GameBeaten)
             ->where('AwardData', $playerGame->game->id);
-        $softcoreBadge = $badge->where('AwardDataExtra', UnlockMode::Softcore);
-        $hardcoreBadge = $badge->where('AwardDataExtra', UnlockMode::Hardcore);
+        $softcoreBadge = (clone $badge)->where('AwardDataExtra', UnlockMode::Softcore);
+        $hardcoreBadge = (clone $badge)->where('AwardDataExtra', UnlockMode::Hardcore);
 
         if ($playerGame->beaten_at === null && $softcoreBadge->exists()) {
             PlayerBadgeLost::dispatch($softcoreBadge->first());
@@ -48,7 +46,7 @@ class RevalidateAchievementSetBadgeEligibility
             $hardcoreBadge->delete();
         }
 
-        if ($playerGame->beaten_at !== null && !$softcoreBadge->exists()) {
+        if ($playerGame->beaten_hardcore_at === null && $playerGame->beaten_at !== null && !$softcoreBadge->exists()) {
             $badge = AddSiteAward(
                 $playerGame->user->username,
                 AwardType::GameBeaten,
@@ -62,6 +60,8 @@ class RevalidateAchievementSetBadgeEligibility
         }
 
         if ($playerGame->beaten_hardcore_at !== null && !$hardcoreBadge->exists()) {
+            $softcoreBadge->delete();
+
             $badge = AddSiteAward(
                 $playerGame->user->username,
                 AwardType::GameBeaten,
@@ -84,20 +84,30 @@ class RevalidateAchievementSetBadgeEligibility
         $badge = $playerGame->user->playerBadges()
             ->where('AwardType', AwardType::Mastery)
             ->where('AwardData', $playerGame->game->id);
-        $softcoreBadge = $badge->where('AwardDataExtra', UnlockMode::Softcore);
-        $hardcoreBadge = $badge->where('AwardDataExtra', UnlockMode::Hardcore);
+        $softcoreBadge = (clone $badge)->where('AwardDataExtra', UnlockMode::Softcore);
+        $hardcoreBadge = (clone $badge)->where('AwardDataExtra', UnlockMode::Hardcore);
 
         if ($playerGame->completed_at === null && $softcoreBadge->exists()) {
-            PlayerBadgeLost::dispatch($softcoreBadge->first());
-            $softcoreBadge->delete();
+            // if the user has at least one unlock for the set, assume there was
+            // a revision and do nothing. if they want to get rid of the badge,
+            // they can reset one or more of the achievements they have.
+            if (!$playerGame->achievements_unlocked && $playerGame->achievements_total) {
+                PlayerBadgeLost::dispatch($softcoreBadge->first());
+                $softcoreBadge->delete();
+            }
         }
 
         if ($playerGame->completed_hardcore_at === null && $hardcoreBadge->exists()) {
-            PlayerBadgeLost::dispatch($hardcoreBadge->first());
-            $hardcoreBadge->delete();
+            // user has no achievements for the set. if the set is empty, assume it
+            // was demoted and keep the badge, otherwise assume they did a full reset
+            // and destroy the badge.
+            if (!$playerGame->achievements_unlocked && !$playerGame->achievements_unlocked_hardcore && $playerGame->achievements_total) {
+                PlayerBadgeLost::dispatch($hardcoreBadge->first());
+                $hardcoreBadge->delete();
+            }
         }
 
-        if ($playerGame->completed_at !== null && !$softcoreBadge->exists()) {
+        if ($playerGame->completed_hardcore_at === null && $playerGame->completed_at !== null && !$softcoreBadge->exists()) {
             $badge = AddSiteAward(
                 $playerGame->user->username,
                 AwardType::Mastery,
@@ -122,6 +132,8 @@ class RevalidateAchievementSetBadgeEligibility
         }
 
         if ($playerGame->completed_hardcore_at !== null && !$hardcoreBadge->exists()) {
+            $softcoreBadge->delete();
+
             $badge = AddSiteAward(
                 $playerGame->user->username,
                 AwardType::Mastery,
@@ -292,18 +304,5 @@ class RevalidateAchievementSetBadgeEligibility
     //
     //         expireGameTopAchievers($game->id);
     //     }
-    // }
-    //
-    // public function playerBadgeExists(string $username, int $awardType, int $data, ?int $dataExtra = null): bool
-    // {
-    //     $badge = PlayerBadge::where('User', $username)
-    //         ->where('AwardType', $awardType)
-    //         ->where('AwardData', $data);
-    //
-    //     if ($dataExtra !== null) {
-    //         $badge->where('AwardDataExtra', $dataExtra);
-    //     }
-    //
-    //     return $badge->exists();
     // }
 }
